+++ blocks/vue-blocks/src/vNextRebateVueApp.vue
@@ -1967,6 +1967,15 @@ function withQueryString(baseUrl) {
           background-size: 1.25rem
         }
 
+        .location-input.is-invalid {
+          outline-color: darkgray;
+          background-color: #fafafa;
+          background-image: url(data:image/svg+xml;base64,PHN2ZyB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciIHZpZXdCb3g9IjAgMCA1MTIgNTEyIj48cGF0aCBmaWxsPSJkYXJrZ3JheSIgb3BhY2l0eT0iLjMiIGQ9Ik0wIDI1NmEyNTYgMjU2IDAgMSAwIDUxMiAwQTI1NiAyNTYgMCAxIDAgMCAyNTZ6TTIzMiAxMjhsNDggMCAwIDI0IDAgMTEyIDAgMjQtNDggMCAwLTI0IDAtMTEyIDAtMjR6bTAgMTkybDQ4IDAgMCA0OC00OCAwIDAtNDh6Ii8+PHBhdGggZmlsbD0iZGFya2dyYXkiIGQ9Ik0yODAgMTUybDAtMjQtNDggMCAwIDI0IDAgMTEyIDAgMjQgNDggMCAwLTI0IDAtMTEyem0wIDE2OGwtNDggMCAwIDQ4IDQ4IDAgMC00OHoiLz48L3N2Zz4=);
+          background-repeat: no-repeat;
+          background-position: right 0.75rem center;
+          background-size: 1.25rem
+        }
+
         .location-input.is-error {
           outline-color: #8b0000;
           background-color: #ffe5e5;